wip: Unifies pinned, selected and receiver video constraints.
This is a WIP effort to unify the pinned, selected and receiver video
constraints messages to the bridge, all while providing full backwards
compatibility with existing client side code.

Phase 2 (implying this WIP is phase 1) is to propagate endpoint
constraints across bridges to enable off-stage video suspension. To
minimize and simplify messaging we consider two types of receivers for a
given stream of a given endpoint, as per @bbaldino's suggestion: a)
endpoint receivers and b) bridge receivers.

The endpoint receivers are straightforward. An endpoint X on bridge B1
receiving 720p from endpoint Y on the same bridge will count as 1
receiver (of 720p from endpoint Y).

The bridge receivers can be thought of this way: Any number of endpoints
on bridge B2 receiving 720p from endpoint Y on bridge B1 will again
count as 1 receiver.